separate bussiness logic to services

+++ internal/campaign/handlers/v1/public_request.go
@@ -1,8 +1,8 @@
 package v1
 
 type DonationRequest struct {
-	Name   string `json:"name" validate:"required,min=3,max=100"`
-	Email  string `json:"email" validate:"required,email"`
-	Amount int    `json:"amount" validate:"required,min=1"`
-	Note   string `json:"note" validate:"omitempty,max=500"`
+	Name   string  `json:"name" validate:"required,min=3,max=100"`
+	Email  string  `json:"email" validate:"required,email"`
+	Amount float32 `json:"amount" validate:"required,min=1"`
+	Note   string  `json:"note" validate:"omitempty,max=500"`
 }
